+++ BeaconSdk/MatrixSdk/Dto/Room/Sync/InvitedRoom.cs
@@ -1,0 +1,7 @@
+namespace MatrixSdk.Dto.Room.Sync
+{
+    public record InvitedRoom(InviteState InviteState)
+    {
+        public InviteState InviteState { get; } = InviteState;
+    }
+}
