+++ BeaconSdk/MatrixSdk/Dto/Room/Sync/LeftRoom.cs
@@ -1,0 +1,8 @@
+namespace MatrixSdk.Dto.Room.Sync
+{
+    public record LeftRoom(TimeLine TimeLine, State State)
+    {
+        public TimeLine TimeLine { get; } = TimeLine;
+        public State State { get; } = State;
+    }
+}
