+++ BeaconSdk/MatrixSdk/Dto/Room/Sync/State.cs
@@ -1,0 +1,10 @@
+namespace MatrixSdk.Dto.Room.Sync
+{
+    using System.Collections.Generic;
+    using Event.State;
+
+    public record State(List<RoomStateEvent> Events)
+    {
+        public List<RoomStateEvent> Events { get; } = Events;
+    }
+}
