+++ BeaconSdk/MatrixSdk/Dto/Room/Sync/TimeLine.cs
@@ -1,0 +1,10 @@
+namespace MatrixSdk.Dto.Room.Sync
+{
+    using System.Collections.Generic;
+    using Event.State;
+
+    public record TimeLine(List<RoomStateEvent> Events)
+    {
+        public List<RoomStateEvent> Events { get; } = Events;
+    }
+}
